Modified gather_tracks() to account for albums that are originally double-sided and thus have a Wikipedia entry with split tracklists.

+++ __main__.py
@@ -10,17 +10,20 @@ import os
 def main(album: str, artist: str) -> None:
     try:
         music_dir = os.path.expanduser('~/Music')
+        download_path = mkdir_album(music_dir, album)
         tracklist_playlist_dict = find_album(album, artist)
         yt_playlist_dict = append_yt_video_to_playlist(tracklist_playlist_dict, artist)
-        download_path = mkdir_album(music_dir, album)
         download_tracks(yt_playlist_dict, download_path)
         print(f"Creation of album \"{album}\" by {artist} complete! Enjoy!")
-    except (KeyboardInterrupt, HTTPError, youtube_dl.utils.DownloadError) as error:
-        print(error)
-        incomp_album_dir = os.listdir(download_path)
-        for file in incomp_album_dir:
-            if file.endswith('.part'):
-                os.remove(os.path.join(incomp_album_dir, file))
+    except (KeyboardInterrupt, HTTPError, youtube_dl.utils.DownloadError):
+        if os.path.exists(download_path):
+            if not os.listdir(download_path):
+                os.remove(download_path)
+            else:
+                incomp_album_dir = os.listdir(download_path)
+                for file in incomp_album_dir:
+                    if file.endswith('.part'):
+                        os.remove(os.path.join(incomp_album_dir, file))
 
 if __name__ == "__main__":
     if len(sys.argv) > 2:
